Intent: Consolidate programming helpers under tnfr.execution Operators involved: [Coupling, Resonance] Affected invariants: [#4, #7] Key changes: - Point CLI helpers and token parsing imports to tnfr.execution - Remove tnfr.program facade and refresh changelog entry - Document new canonical module for programmatic helpers Expected risks/dissonances: Legacy imports from tnfr.program will fail; mitigated by explicit changelog guidance Metrics: (C(t), Si, νf, phase) unaffected; regression covered by CLI/token parser tests Equivalence map: tnfr.program.{play,seq,block,wait,target} -> tnfr.execution.{play,seq,block,wait,target}
