+++ README.md
@@ -278,7 +278,7 @@ https://chatgpt.com/g/g-67abc78885a88191b2d67f94fd60dc97-tnfr-teoria-de-la-natur
 ## Changelog
 
 * Removed deprecated alias `sigma_vector_global`; use `sigma_vector_from_graph` instead.
-* Cleaned up `tnfr.program.__all__` to exclude private helpers.
+* Removed legacy `tnfr.program` alias; import programming helpers from `tnfr.execution`.
 * Stopped re-exporting ``CallbackSpec`` and ``apply_topological_remesh`` at the
   package root; import them via ``tnfr.trace`` and ``tnfr.operators``.
